backend/src/yasg.py
from django.urls import path
from drf_yasg import openapi
from drf_yasg.views import get_schema_view
from rest_framework import permissions
from decouple import config


schema_view = get_schema_view(
    openapi.Info(
        title="Shop API",
        default_version='v1',
        description="Shop",
        terms_of_service="https://www.google.com/policies/terms/",
        license=openapi.License(name="BSD License"),
    ),
    public=True,
    url=config('EMAIL_URL'),
    permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path('api/swagger/', schema_view.with_ui('swagger',
                                             cache_timeout=0), name='schema-swagger-ui'),

    path('api/api.json/', schema_view.without_ui(cache_timeout=0),
         name='schema-swagger-ui'),
    path('api/redoc/', schema_view.with_ui('redoc',
                                           cache_timeout=0), name='schema-redoc'),
]
